refactor: extracted utils for expandable data table.
